Rename internal compose packages
