+++ app/src/main/kotlin/it/vfsfitvnm/vimusic/ui/screens/home/HomeScreen.kt
@@ -6,13 +6,13 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.saveable.rememberSaveableStateHolder
 import androidx.compose.ui.platform.LocalContext
 import it.vfsfitvnm.compose.persist.PersistMapCleanup
-import it.vfsfitvnm.route.RouteHandler
-import it.vfsfitvnm.route.defaultStacking
-import it.vfsfitvnm.route.defaultStill
-import it.vfsfitvnm.route.defaultUnstacking
-import it.vfsfitvnm.route.isStacking
-import it.vfsfitvnm.route.isUnknown
-import it.vfsfitvnm.route.isUnstacking
+import it.vfsfitvnm.compose.routing.RouteHandler
+import it.vfsfitvnm.compose.routing.defaultStacking
+import it.vfsfitvnm.compose.routing.defaultStill
+import it.vfsfitvnm.compose.routing.defaultUnstacking
+import it.vfsfitvnm.compose.routing.isStacking
+import it.vfsfitvnm.compose.routing.isUnknown
+import it.vfsfitvnm.compose.routing.isUnstacking
 import it.vfsfitvnm.vimusic.Database
 import it.vfsfitvnm.vimusic.R
 import it.vfsfitvnm.vimusic.models.SearchQuery
