+++ compose-routing/src/main/kotlin/it/vfsfitvnm/compose/routing/GlobalRoute.kt
@@ -1,4 +1,4 @@
-package it.vfsfitvnm.route
+package it.vfsfitvnm.compose.routing
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
