+++ compose-routing/src/main/kotlin/it/vfsfitvnm/compose/routing/RouteHandler.kt
@@ -1,4 +1,4 @@
-package it.vfsfitvnm.route
+package it.vfsfitvnm.compose.routing
 
 import androidx.activity.compose.BackHandler
 import androidx.activity.compose.LocalOnBackPressedDispatcherOwner
